delete excess code and modify thread pool
