+++ tests/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory(opt)
 add_subdirectory(thread)
-add_subdirectory(test_server)
-add_subdirectory(test_client)
+#add_subdirectory(test_server)
+#add_subdirectory(test_client)
